type, interface basics
Create type square, circle, attach method area() to each.

Create a type shape interface as anything which has the area method.

Create a func info which takes type shape and then prints the area.

Create a value of type square and a value of type circle.

Use func info to print the area of both.